switch integration tests to typescript
