+++ tsconfig.tests.json
@@ -6,13 +6,12 @@
     "target": "es6",
     "sourceMap": false,
     "outDir": "dist",
-    "types": ["node"]
+    "types": ["mocha", "node"]
   },
   "include": [
-    "**/*.ts"
+    "test/**/*.ts"
   ],
   "exclude": [
-    "node_modules",
-    "test"
+    "node_modules"
   ]
 }
