+++ tslint.json
@@ -1,3 +1,6 @@
 {
-    "extends": "tslint:recommended"
+    "extends": "tslint:recommended",
+    "rules": {
+        "no-console": false
+    }
 }
